# ArchitectureRefactoringHelper/Repository/Models/Database/ToolSource.cs
using System.ComponentModel.DataAnnotations;
using System.ComponentModel.DataAnnotations.Schema;
using System.Text.Json.Serialization;

namespace Repository.Models.Database;

[Table(Constants.TableNameToolSource)]
public class ToolSource
{
    [Key]
    [JsonPropertyName("toolSourceId")]
    public int ToolSourceId { get; set; }
    
    [Required]
    [JsonPropertyName("name")]
    public string Name { get; set; }

    [Required]
    [JsonPropertyName("author")]
    public string Author { get; set; }

    [JsonPropertyName("link")]
    public string? Link { get; set; }

    [JsonPropertyName("description")]
    public string? Description { get; set; }

    [JsonIgnore]
    public Tool? Tool { get; set; }
}