# ArchitectureRefactoringHelper/Repository/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Repository, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/tools": {
      "get": {
        "tags": [
          "Tool"
        ],
        "summary": "Receives a complete list of all Tool items",
        "operationId": "ListTools",
        "parameters": [
          {
            "name": "withDetails",
            "in": "query",
            "description": "Decides whether the tools are returned with all details or not",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tool"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tool"
        ],
        "summary": "Creates a new Tool.\r\nToolTypes that don't exist will be created in the process.",
        "operationId": "AddTool",
        "requestBody": {
          "description": "The Tool we want to add",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Tool"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Tool"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Tool"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns the newly created Tool",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tool"
                }
              }
            }
          },
          "400": {
            "description": "If the Tool is null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "If a Tool with the same title already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tools/{id}": {
      "get": {
        "tags": [
          "Tool"
        ],
        "summary": "Receives a Tool",
        "operationId": "GetTool",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tool"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tool"
        ],
        "summary": "Deletes a Tool",
        "operationId": "DeleteTool",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/tools/{id}/tool-types": {
      "post": {
        "tags": [
          "Tool"
        ],
        "operationId": "AddToolTypeToTool",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ToolType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/tools/{id}/existing-cards": {
      "delete": {
        "tags": [
          "Tool"
        ],
        "operationId": "RemoveToolExistingCards",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/tools/tool-types": {
      "get": {
        "tags": [
          "Tool"
        ],
        "operationId": "ListToolTypes",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ToolType"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tool"
        ],
        "operationId": "AddToolType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ToolType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolType"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tools/tool-types/{name}": {
      "delete": {
        "tags": [
          "Tool"
        ],
        "operationId": "DeleteToolType",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AccuracyPrecision": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "AnalysisType": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ApproachOutput": {
        "required": [
          "architecture",
          "serviceType"
        ],
        "type": "object",
        "properties": {
          "approachOutputId": {
            "type": "integer",
            "format": "int32"
          },
          "architecture": {
            "$ref": "#/components/schemas/Architecture"
          },
          "serviceType": {
            "$ref": "#/components/schemas/ServiceType"
          }
        },
        "additionalProperties": false
      },
      "ApproachProcess": {
        "type": "object",
        "properties": {
          "approachProcessId": {
            "type": "integer",
            "format": "int32"
          },
          "qualities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Quality"
            },
            "nullable": true
          },
          "qualitySublevels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QualitySublevel"
            },
            "nullable": true
          },
          "directions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Direction"
            },
            "nullable": true
          },
          "automationLevels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AutomationLevel"
            },
            "nullable": true
          },
          "analysisTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnalysisType"
            },
            "nullable": true
          },
          "techniques": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Technique"
            },
            "nullable": true
          },
          "processStrategies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProcessStrategy"
            },
            "nullable": true
          },
          "atomarUnits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AtomarUnit"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ApproachSource": {
        "required": [
          "authors",
          "title",
          "year"
        ],
        "type": "object",
        "properties": {
          "approachSourceId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "link": {
            "type": "string",
            "nullable": true
          },
          "authors": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ApproachUsability": {
        "required": [
          "accuracyPrecision",
          "validationMethod"
        ],
        "type": "object",
        "properties": {
          "approachUsabilityId": {
            "type": "integer",
            "format": "int32"
          },
          "noToolSupport": {
            "type": "boolean",
            "nullable": true
          },
          "refactoringApproach": {
            "$ref": "#/components/schemas/RefactoringApproach"
          },
          "resultsQuality": {
            "$ref": "#/components/schemas/ResultsQuality"
          },
          "toolSupport": {
            "$ref": "#/components/schemas/ToolSupport"
          },
          "accuracyPrecision": {
            "$ref": "#/components/schemas/AccuracyPrecision"
          },
          "validationMethod": {
            "$ref": "#/components/schemas/ValidationMethod"
          },
          "tools": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tool"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Architecture": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "AtomarUnit": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "AutomationLevel": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CalculationMetric": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Direction": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DomainArtifactInput": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ExecutableInput": {
        "required": [
          "language",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ModelArtifactInput": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "ProcessStrategy": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Quality": {
        "required": [
          "category",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/QualityCategory"
          },
          "qualitySublevels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QualitySublevel"
            },
            "nullable": true
          },
          "qualityTradeOffs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Quality"
            },
            "nullable": true
          },
          "calculationMetrics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CalculationMetric"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "QualityCategory": {
        "enum": [
          "Attribute",
          "SystemProperty"
        ],
        "type": "string"
      },
...
      "QualitySublevel": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "qualityName": {
            "type": "string",
            "nullable": true
          },
          "qualityTradeOffs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Quality"
            },
            "nullable": true
          },
          "calculationMetrics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CalculationMetric"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RefactoringApproach": {
        "required": [
          "identifier"
        ],
        "type": "object",
        "properties": {
          "refactoringApproachId": {
            "type": "integer",
            "format": "int32"
          },
          "identifier": {
            "type": "string"
          },
          "approachSource": {
            "$ref": "#/components/schemas/ApproachSource"
          },
          "domainArtifactInputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DomainArtifactInput"
            },
            "nullable": true
          },
          "runtimeArtifactInputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RuntimeArtifactInput"
            },
            "nullable": true
          },
          "modelArtifactInputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelArtifactInput"
            },
            "nullable": true
          },
          "executableInputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExecutableInput"
            },
            "nullable": true
          },
          "approachProcess": {
            "$ref": "#/components/schemas/ApproachProcess"
          },
          "approachOutputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApproachOutput"
            },
            "nullable": true
          },
          "representationOutputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Representation"
            },
            "nullable": true
          },
          "approachUsability": {
            "$ref": "#/components/schemas/ApproachUsability"
          }
        },
        "additionalProperties": false
      },
      "Representation": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ResultsQuality": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RuntimeArtifactInput": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ServiceType": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Technique": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Tool": {
        "required": [
          "identifier"
        ],
        "type": "object",
        "properties": {
          "toolId": {
            "type": "integer",
            "format": "int32"
          },
          "identifier": {
            "type": "string"
          },
          "toolSource": {
            "$ref": "#/components/schemas/ToolSource"
          },
          "toolTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ToolType"
            },
            "nullable": true
          },
          "approachUsabilities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApproachUsability"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ToolSource": {
        "required": [
          "author",
          "name"
        ],
        "type": "object",
        "properties": {
          "toolSourceId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "link": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ToolSupport": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ToolType": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ValidationMethod": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}